+++ lib/src/converter/svg/svg_exporter.dart
@@ -6,12 +6,13 @@ import 'package:xml/xml.dart';
 /// A utility class for decoding a [SketchController] to SVG data
 class SvgExporter implements Exporter {
   @override
-  String export(SketchController controller) {
+  String export(SketchController controller, {bool exportBackgroundColor = false}) {
     final builder = XmlBuilder();
     builder.processing('xml', 'version="1.0"');
     builder.element("svg", nest: () {
-      builder.attribute("viewport-fill",
-          _flutterColorToSvgColor(controller.backgroundColor.value));
+      if (exportBackgroundColor) {
+        builder.attribute("viewport-fill", _flutterColorToSvgColor(controller.backgroundColor.value));
+      }
       for (var layer in controller.layers) {
         for (var stroke in layer.painter.strokes) {
           _toPath(builder, stroke);
@@ -22,8 +23,7 @@ class SvgExporter implements Exporter {
   }
 
   void _toPath(XmlBuilder builder, Stroke stroke) {
-    var d =
-        "M${stroke.points.first.dx.toInt()} ${stroke.points.first.dy.toInt()}";
+    var d = "M${stroke.points.first.dx.toInt()} ${stroke.points.first.dy.toInt()}";
     for (var point in stroke.points.skip(1)) {
       d += " L${point.dx.toInt()} ${point.dy.toInt()}";
     }