+++ lib/src/converter/svg/svg_importer.dart
@@ -14,8 +14,7 @@ class SvgImporter implements Importer {
     for (var pathNode in document.rootElement.children) {
       final d = pathNode.getAttribute("d")!;
       final stroke = pathNode.getAttribute("stroke")!;
-      final strokeOpacity =
-          double.parse(pathNode.getAttribute("stroke-opacity")!);
+      final strokeOpacity = double.parse(pathNode.getAttribute("stroke-opacity")!);
       final strokeWidth = double.parse(pathNode.getAttribute("stroke-width")!);
 
       final commands = d.split(" ");
@@ -28,14 +27,11 @@ class SvgImporter implements Importer {
         points.add(Offset(dx.toDouble(), dy.toDouble()));
       }
 
-      final strokeColor =
-          _svgColorToFlutterColor(stroke).withOpacity(strokeOpacity);
+      final strokeColor = _svgColorToFlutterColor(stroke).withOpacity(strokeOpacity);
       strokes.add(Stroke(points, strokeColor, strokeWidth));
     }
     final viewportFill = document.rootElement.getAttribute("viewport-fill");
-    final backgroundColor = viewportFill != null
-        ? _svgColorToFlutterColor(viewportFill)
-        : Colors.lightGreenAccent;
+    final backgroundColor = viewportFill != null ? _svgColorToFlutterColor(viewportFill) : Colors.transparent;
     controller.init(strokes, backgroundColor);
   }