+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.1.3
+
+* added the weight to active tool style
+
+## 1.1.2
+
+* made the background color optional for importer
+
+## 1.1.1
+
+* made the background color optional for exporter
+
 ## 1.1.0
 
 * opted into the null safety